chore(button): only include perf stories when env variable is present
