+++ .github/workflows/on_pull_request.yml
@@ -359,6 +359,7 @@ jobs:
           exitOnceUploaded: true
         env:
           STORYBOOK_GITHUB_SHA: ${{ github.event.pull_request.head.sha }}
+          INCLUDE_PERF_STORIES: true
 
   pr-categorizer:
     name: Categorize the PR using labels
