--- packages/paste-core/components/button/stories/performance/tests/partials.tsx
@@ -1,17 +1,0 @@
-import * as React from 'react';
-import {Box} from '@twilio-paste/box';
-
-export const Wrapper: React.FC = ({children}) => (
-  <Box
-    padding="space30"
-    display="flex"
-    alignItems="baseline"
-    justifyContent="space-between"
-    maxWidth="50%"
-    columnGap="space80"
-    verticalAlign="center"
-  >
-    {children}
-  </Box>
-);
-Wrapper.displayName = 'PERF_TEST_WRAPPER';
