+++ packages/paste-core/components/button/stories/performance/third-test.stories.performance.tsx
@@ -1,12 +1,14 @@
 import * as React from 'react';
 import {Box} from '@twilio-paste/box';
 import {useSpring, animated} from '@twilio-paste/animation-library';
+// eslint-disable-next-line import/no-extraneous-dependencies
+import {Title, Description, Primary} from '@storybook/addon-docs';
 import {ArrowForwardIcon} from '@twilio-paste/icons/esm/ArrowForwardIcon';
 import type {StoryFn} from '@storybook/react';
-import type {ButtonProps, ButtonStates} from '../../../src/types';
+import type {ButtonProps, ButtonStates} from '../../src/types';
 
-import {Wrapper} from './partials';
-import {getButtonComponent} from './helpers';
+import {getButtonComponent, StoryWrapper} from './helpers-and-partials';
+import type {DecoratorStory} from './types';
 
 const getButtonState = (disabled?: boolean, loading?: boolean): ButtonStates => {
   if (disabled) {
@@ -17,6 +19,7 @@ const getButtonState = (disabled?: boolean, loading?: boolean): ButtonStates => 
   }
   return 'default';
 };
+
 const AnimatedBox = animated(Box);
 
 const variantsWithoutBoundingBox = new Set(['link', 'destructive_link', 'inverse_link', 'reset']);
@@ -83,38 +86,95 @@ Button.displayName = 'Button';
 
 // ========================== Stories ============================
 
+const CODE = `const Button = React.forwardRef<HTMLButtonElement, ButtonProps>(
+  ({element = 'BUTTON', size, variant, children, disabled, loading, onMouseEnter, onMouseLeave, ...props}, ref) => {
+    const buttonState = getButtonState(disabled, loading);
+    const [hovered, setHovered] = React.useState(false);
+    const arrowIconStyles = useSpring({
+      translateX: hovered ? '4px' : '0px',
+      config: {
+        mass: 0.1,
+        tension: 275,
+        friction: 16,
+      },
+    });
+
+    const ButtonComponent = getButtonComponent(variant);
+
+    return (
+      <ButtonComponent
+        {...props}
+        onMouseEnter={(event) => {
+          if (typeof onMouseEnter === 'function') {
+            onMouseEnter(event);
+          }
+          setHovered(true);
+        }}
+        onMouseLeave={(event) => {
+          if (typeof onMouseLeave === 'function') {
+            onMouseLeave(event);
+          }
+          setHovered(false);
+        }}
+        buttonState={buttonState}
+        disabled={buttonState !== 'default'}
+        element={element}
+        variant={variant}
+        size="default"
+        aria-busy={buttonState === 'loading' ? 'true' : 'false'}
+        className={undefined}
+        style={undefined}
+        ref={ref}
+      >
+        <Box
+          as="span"
+          display="flex"
+          textDecoration="inherit"
+          justifyContent={variant && variantsWithoutBoundingBox.has(variant) ? null : 'center'}
+          columnGap="space20"
+          alignItems="center"
+        >
+          {children}
+          <AnimatedBox style={arrowIconStyles}>
+            <ArrowForwardIcon decorative />
+          </AnimatedBox>
+        </Box>
+      </ButtonComponent>
+    );
+  }
+);`;
+
 // eslint-disable-next-line import/no-default-export
 export default {
   title: 'Button-Performance/Tests',
   component: Button,
   decorators: [
-    (Story: StoryFn) => {
-      return (
-        <Wrapper>
-          <div>
-            <h2>Test 3: Button with variant selection function</h2>
-            <p>
-              Button stripped down to remove:
-              <ul>
-                <li>loading state</li>
-                <li>disabled state</li>
-                <li>smart sizing</li>
-              </ul>
-              Other changes include:
-              <ul>
-                <li>flattened render</li>
-                <li>removal of state from component</li>
-                <li>animation added back</li>
-              </ul>
-            </p>
-          </div>
-          <div>
-            <Story />
-          </div>
-        </Wrapper>
-      );
-    },
+    (Story: DecoratorStory) => (
+      <StoryWrapper>
+        <Story />
+      </StoryWrapper>
+    ),
   ],
+  parameters: {
+    docs: {
+      description: {
+        component: 'Test 2 button with conditional rendering of different button variants added back.',
+      },
+      source: {
+        state: 'open',
+
+        code: CODE,
+        language: 'jsx',
+      },
+      page: () => (
+        <>
+          <Title>Test 3: Button with variant selection function</Title>
+          <Description />
+          <Primary />
+        </>
+      ),
+    },
+  },
 };
 
 export const Third: StoryFn = () => <Button variant="primary">Test</Button>;
