+++ packages/paste-core/components/button/stories/performance/types.ts
@@ -1,0 +1,3 @@
+import type {DecoratorFn} from '@storybook/react';
+
+export type DecoratorStory = Parameters<DecoratorFn>[0];
